This probes whether environment variables for a container can be determined by setting them on the docker-cli
The answer is no - just set them on the command you're running inside the container if you want this.
This makes _sense_ the environment of the docker-cli is different than the environment of the container! This kind of isolation is the core of docker!

But it's nice that we can _confirm_ that this is indeed the behaviour very quickly.